REF: share isna, nbytes
